+++ src/index.js
@@ -1,15 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { createHashHistory } from "history";
-import { Router, Route, Switch } from "react-router-dom";
+import { Router, Route, Switch, Redirect } from "react-router-dom";
 import ReactGA from "react-ga";
 import ReactPixel from "react-facebook-pixel";
 
 import "bootstrap/dist/css/bootstrap.css";
-import "assets/scss/now-ui-dashboard.css";
+import "assets/scss/now-ui-dashboard.scss?v1.2.0";
 import "assets/css/demo.css";
 
-import indexRoutes from "routes/index.jsx";
+import AdminLayout from "layouts/Admin.jsx";
 
 const hist = createHashHistory();
 
@@ -32,9 +32,8 @@ hist.listen(location => {
 ReactDOM.render(
   <Router history={hist}>
     <Switch>
-      {indexRoutes.map((prop, key) => {
-        return <Route path={prop.path} key={key} component={prop.component} />;
-      })}
+      <Route path="/admin" render={props => <AdminLayout {...props} />} />
+      <Redirect to="/admin/dashboard" />
     </Switch>
   </Router>,
   document.getElementById("root")