+++ src/layouts/Admin.jsx
@@ -1,0 +1,61 @@
+import React from "react";
+// javascript plugin used to create scrollbars on windows
+import PerfectScrollbar from "perfect-scrollbar";
+
+// reactstrap components
+import { Route, Switch } from "react-router-dom";
+
+// core components
+import DemoNavbar from "components/Navbars/DemoNavbar.jsx";
+import Footer from "components/Footer/Footer.jsx";
+import Sidebar from "components/Sidebar/Sidebar.jsx";
+
+import routes from "routes.js";
+
+var ps;
+
+class Dashboard extends React.Component {
+  mainPanel = React.createRef();
+  componentDidMount() {
+    if (navigator.platform.indexOf("Win") > -1) {
+      ps = new PerfectScrollbar(this.mainPanel.current);
+      document.body.classList.toggle("perfect-scrollbar-on");
+    }
+  }
+  componentWillUnmount() {
+    if (navigator.platform.indexOf("Win") > -1) {
+      ps.destroy();
+      document.body.classList.toggle("perfect-scrollbar-on");
+    }
+  }
+  componentDidUpdate(e) {
+    if (e.history.action === "PUSH") {
+      this.mainPanel.current.scrollTop = 0;
+      document.scrollingElement.scrollTop = 0;
+    }
+  }
+  render() {
+    return (
+      <div className="wrapper">
+        <Sidebar {...this.props} routes={routes} />
+        <div className="main-panel" ref={this.mainPanel}>
+          <DemoNavbar {...this.props} />
+          <Switch>
+            {routes.map((prop, key) => {
+              return (
+                <Route
+                  path={prop.layout + prop.path}
+                  component={prop.component}
+                  key={key}
+                />
+              );
+            })}
+          </Switch>
+          <Footer fluid />
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Dashboard;