+++ src/routes.js
@@ -1,0 +1,69 @@
+import Dashboard from "views/Dashboard.jsx";
+import Notifications from "views/Notifications.jsx";
+import Icons from "views/Icons.jsx";
+import Typography from "views/Typography.jsx";
+import TableList from "views/TableList.jsx";
+import Maps from "views/Maps.jsx";
+import Upgrade from "views/Upgrade.jsx";
+import UserPage from "views/UserPage.jsx";
+
+var dashRoutes = [
+  {
+    path: "/dashboard",
+    name: "Dashboard",
+    icon: "design_app",
+    component: Dashboard,
+    layout: "/admin"
+  },
+  {
+    path: "/icons",
+    name: "Icons",
+    icon: "design_image",
+    component: Icons,
+    layout: "/admin"
+  },
+  {
+    path: "/maps",
+    name: "Maps",
+    icon: "location_map-big",
+    component: Maps,
+    layout: "/admin"
+  },
+  {
+    path: "/notifications",
+    name: "Notifications",
+    icon: "ui-1_bell-53",
+    component: Notifications,
+    layout: "/admin"
+  },
+  {
+    path: "/user-page",
+    name: "User Profile",
+    icon: "users_single-02",
+    component: UserPage,
+    layout: "/admin"
+  },
+  {
+    path: "/extended-tables",
+    name: "Table List",
+    icon: "files_paper",
+    component: TableList,
+    layout: "/admin"
+  },
+  {
+    path: "/typography",
+    name: "Typography",
+    icon: "design-2_ruler-pencil",
+    component: Typography,
+    layout: "/admin"
+  },
+  {
+    pro: true,
+    path: "/upgrade",
+    name: "Upgrade to PRO",
+    icon: "objects_spaceship",
+    component: Upgrade,
+    layout: "/admin"
+  }
+];
+export default dashRoutes;